Improve error reporting during indexing
The previous behavior was to basically suppress all error diagnostics
generated by Clang during indexing, and only rely on the return value of
`ClangTool::run` to decide whether a file was parsed correctly or not.

For irrecoverable errors such as missing files this doesn't give one
much to work with. The new behavior is to parse all diagnostics and
check for missing files, and if there are some, print those immediately.

For other errors that may be encountered, we now keep count of how many
there are, and offer the option to print them using --verbose output.